requirements.txt

alembic==1.6.5
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Migrate==3.0.1
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
mysql==0.0.3
mysqlclient==2.0.3
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.4.17
Werkzeug==2.0.1